+++ src/Symfony/Component/Security/Core/Authentication/Token/AbstractToken.php
@@ -133,20 +133,67 @@ abstract class AbstractToken implements TokenInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @final since Symfony 4.3, use getState() instead
+     *
+     * @internal since Symfony 4.3, use getState() instead
      */
     public function serialize()
     {
-        $serialized = [$this->user, $this->authenticated, $this->roles, $this->attributes];
-
-        return $this->doSerialize($serialized, \func_num_args() ? \func_get_arg(0) : null);
+        return $this->doSerialize($this->getState(), \func_num_args() ? \func_get_arg(0) : null);
     }
 
     /**
      * {@inheritdoc}
+     *
+     * @final since Symfony 4.3, use setState() instead
+     *
+     * @internal since Symfony 4.3, use setState() instead
      */
     public function unserialize($serialized)
     {
-        list($this->user, $this->authenticated, $this->roles, $this->attributes) = \is_array($serialized) ? $serialized : unserialize($serialized);
+        $this->setState(\is_array($serialized) ? $serialized : unserialize($serialized));
+    }
+
+    /**
+     * Returns all the necessary state of the object for serialization purposes.
+     *
+     * There is no need to serialize any entry, they should be returned as-is.
+     * If you extend this method, keep in mind you MUST guarantee parent data is present in the state.
+     * Here is an example of how to extend this method:
+     * <code>
+     *     protected function getState(): array
+     *     {
+     *         return [$this->childAttribute, parent::getState()];
+     *     }
+     * </code>
+     *
+     * @see setState()
+     */
+    protected function getState(): array
+    {
+        return [$this->user, $this->authenticated, $this->roles, $this->attributes];
+    }
+
+    /**
+     * Restores the object state from an array given by getState().
+     *
+     * There is no need to unserialize any entry in $data, they are already ready-to-use.
+     * If you extend this method, keep in mind you MUST pass the parent data to its respective class.
+     * Here is an example of how to extend this method:
+     * <code>
+     *     protected function setState(array $data)
+     *     {
+     *         [$this->childAttribute, $parentData] = $data;
+     *         parent::setState($parentData);
+     *     }
+     * </code>
+     *
+     * @see getState()
+     */
+    protected function setState(array $data)
+    {
+        [$this->user, $this->authenticated, $this->roles, $this->attributes] = $data;
     }
 
     /**